week1/bebamotiend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
stripe==6.6.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
python-dateutil==2.8.2